Add parameter tests to arguments_data_test.go

+++ function/float64_parameter_validator.go
@@ -17,7 +17,7 @@ type Float64Validator interface {
 type Float64Request struct {
 	// ArgumentPosition contains the position of the argument for validation.
 	// Use this position for any response diagnostics.
-	ArgumentPosition int
+	ArgumentPosition int64
 
 	// Value contains the value of the argument for validation.
 	Value types.Float64
